requirements.txt
ruamel.yaml>=0.17.32
